+++ R/z-score-weight-for-lenhei.R
@@ -15,9 +15,18 @@
 #' @include z-score-helper.R
 #' @noRd
 anthro_zscore_weight_for_lenhei <-
-  function(weight, lenhei, lenhei_unit, age_in_days, age_in_months, sex, oedema,
-           flag_threshold = 5, growthstandards_wfl = growthstandards_wflanthro,
-           growthstandards_wfh = growthstandards_wfhanthro) {
+  function(
+    weight,
+    lenhei,
+    lenhei_unit,
+    age_in_days,
+    age_in_months,
+    sex,
+    oedema,
+    flag_threshold = 5,
+    growthstandards_wfl = growthstandards_wflanthro,
+    growthstandards_wfh = growthstandards_wfhanthro
+  ) {
     stopifnot(is.numeric(weight))
     stopifnot(is.numeric(weight))
     stopifnot(is.character(oedema) && all(oedema %in% c("y", "n")))
@@ -31,10 +40,6 @@ anthro_zscore_weight_for_lenhei <-
     # join that with the growthstandards
     # then we have everything to compute the zscores
     n <- length(lenhei)
-
-    # clean weight/lenhei
-    weight[weight < 0.9 | weight > 58.0] <- NA_real_
-    lenhei[lenhei < 38.0 | lenhei > 150.0] <- NA_real_
 
     # we also need to interpolate lenhei under certain coniditions
     low_lenhei <- trunc(lenhei * 10) / 10
@@ -53,55 +58,71 @@ anthro_zscore_weight_for_lenhei <-
         lenhei_unit == "l") |
       (is.na(age_in_days) &
         is.na(lenhei_unit) &
-        !is.na(lenhei) & lenhei < 87))
+        !is.na(lenhei) &
+        lenhei < 87))
 
     join_on_h <- ((!is.na(age_in_days) & age_in_days >= 731) |
       (is.na(age_in_days) &
         !is.na(lenhei_unit) &
         lenhei_unit == "h") |
-      (is.na(age_in_days) & is.na(lenhei_unit) &
-        !is.na(lenhei) & lenhei >= 87))
+      (is.na(age_in_days) & is.na(lenhei_unit) & !is.na(lenhei) & lenhei >= 87))
 
-    input_df <- data.frame(weight, sex, lenhei_unit, low_lenhei, upp_lenhei,
+    input_df <- data.frame(
+      weight,
+      sex,
+      lenhei_unit,
+      low_lenhei,
+      upp_lenhei,
       diff_lenhei,
       ordering = seq_len(n),
-      join_col = ifelse(join_on_l, rep.int("l", n),
-        ifelse(join_on_h,
-          rep.int("h", n),
-          rep.int(NA_character_, n)
-        )
+      join_col = ifelse(
+        join_on_l,
+        rep.int("l", n),
+        ifelse(join_on_h, rep.int("h", n), rep.int(NA_character_, n))
       )
     )
-    merged_df <- merge(input_df,
+    merged_df <- merge(
+      input_df,
       growthstandards,
       by.x = c("sex", "low_lenhei", "join_col"),
       by.y = c("sex", "lenhei", "lorh"),
-      all.x = TRUE, sort = FALSE, suffixes = c("", "_lower")
+      all.x = TRUE,
+      sort = FALSE,
+      suffixes = c("", "_lower")
     )
-    merged_df <- merge(merged_df,
+    merged_df <- merge(
+      merged_df,
       growthstandards,
       by.x = c("sex", "upp_lenhei", "join_col"),
       by.y = c("sex", "lenhei", "lorh"),
-      all.x = TRUE, sort = FALSE, suffixes = c("", "_upper")
+      all.x = TRUE,
+      sort = FALSE,
+      suffixes = c("", "_upper")
     )
     merged_df <- merged_df[order(merged_df$ordering), ]
 
     y <- merged_df[["weight"]]
-    m <- ifelse(diff_lenhei > 0,
-      merged_df[["m"]] + diff_lenhei *
-        (merged_df[["m_upper"]] - merged_df[["m"]]),
+    m <- ifelse(
+      diff_lenhei > 0,
+      merged_df[["m"]] +
+        diff_lenhei *
+          (merged_df[["m_upper"]] - merged_df[["m"]]),
       merged_df[["m"]]
     )
     m <- as.numeric(m)
-    l <- ifelse(diff_lenhei > 0,
-      merged_df[["l"]] + diff_lenhei *
-        (merged_df[["l_upper"]] - merged_df[["l"]]),
+    l <- ifelse(
+      diff_lenhei > 0,
+      merged_df[["l"]] +
+        diff_lenhei *
+          (merged_df[["l_upper"]] - merged_df[["l"]]),
       merged_df[["l"]]
     )
     l <- as.numeric(l)
-    s <- ifelse(diff_lenhei > 0,
-      merged_df[["s"]] + diff_lenhei *
-        (merged_df[["s_upper"]] - merged_df[["s"]]),
+    s <- ifelse(
+      diff_lenhei > 0,
+      merged_df[["s"]] +
+        diff_lenhei *
+          (merged_df[["s_upper"]] - merged_df[["s"]]),
       merged_df[["s"]]
     )
     s <- as.numeric(s)
@@ -110,12 +131,10 @@ anthro_zscore_weight_for_lenhei <-
     zscore <- round(zscore, digits = 2L)
 
     valid_zscore <- !is.na(lenhei) &
-      ifelse(join_on_l,
+      ifelse(
+        join_on_l,
         lenhei >= 45 & lenhei <= 110,
-        ifelse(join_on_h,
-          lenhei >= 65 & lenhei <= 120,
-          rep.int(FALSE, n)
-        )
+        ifelse(join_on_h, lenhei >= 65 & lenhei <= 120, rep.int(FALSE, n))
       )
 
     valid_zscore <- valid_zscore & !(oedema %in% "y")