+++ data-raw/growthstandards/growthstandards.R
@@ -28,7 +28,8 @@ growthstandards_wfhanthro <- as.data.frame(
 )
 growthstandards_wfhanthro$sex <- as.integer(growthstandards_wfhanthro$sex)
 
-usethis::use_data(growthstandards_weianthro,
+usethis::use_data(
+  growthstandards_weianthro,
   growthstandards_lenanthro,
   growthstandards_bmianthro,
   growthstandards_wflanthro,
@@ -37,5 +38,7 @@ usethis::use_data(growthstandards_weianthro,
   growthstandards_acanthro,
   growthstandards_tsanthro,
   growthstandards_ssanthro,
-  overwrite = TRUE, internal = TRUE
+  overwrite = TRUE,
+  internal = TRUE,
+  version = 2 # we set it to version 2 to be able to support older R versions
 )