+++ tests/testthat/test-zscore-bmi-for-age.R
@@ -1,37 +1,71 @@
 context("zscores - bmi for age")
 describe("anthro_zscore_bmi_for_age", {
   it("fails if bmi is not numeric", {
-    expect_error(anthro_zscore_bmi_for_age(
-      "50", 44,
-      to_months(44), 1, "n"
-    ), "measure")
+    expect_error(
+      anthro_zscore_bmi_for_age(
+        "50",
+        44,
+        to_months(44),
+        1,
+        "n"
+      ),
+      "measure"
+    )
   })
   it("fails if age is not numeric", {
-    expect_error(anthro_zscore_bmi_for_age(
-      50, "44",
-      to_months(44), 1, "n"
-    ), "age")
+    expect_error(
+      anthro_zscore_bmi_for_age(
+        50,
+        "44",
+        to_months(44),
+        1,
+        "n"
+      ),
+      "age"
+    )
   })
   it("fails if sex is not 1 or 2", {
-    expect_error(anthro_zscore_bmi_for_age(
-      50, 44,
-      to_months(44), 3, "n"
-    ), "sex")
-    expect_error(anthro_zscore_bmi_for_age(
-      50, 44,
-      to_months(44), "F", "n"
-    ), "sex")
+    expect_error(
+      anthro_zscore_bmi_for_age(
+        50,
+        44,
+        to_months(44),
+        3,
+        "n"
+      ),
+      "sex"
+    )
+    expect_error(
+      anthro_zscore_bmi_for_age(
+        50,
+        44,
+        to_months(44),
+        "F",
+        "n"
+      ),
+      "sex"
+    )
   })
   it("sets result to NA if oedema is y", {
-    expect_true(is.na(anthro_zscore_bmi_for_age(
-      50, 44,
-      to_months(44), 1, "y"
-    )$zbmi))
+    expect_true(is.na(
+      anthro_zscore_bmi_for_age(
+        50,
+        44,
+        to_months(44),
+        1,
+        "y"
+      )$zbmi
+    ))
   })
   it("has a default flag threshold of -6;5", {
     observed <- anthro_zscore_bmi_for_age(25, 1, to_months(1), 2, "n")
     expect_equal(observed$fbmi, 1L)
-    observed <- anthro_zscore_bmi_for_age(10.37, 1, to_months(1), 2, "n",
+    observed <- anthro_zscore_bmi_for_age(
+      10.37,
+      1,
+      to_months(1),
+      2,
+      "n",
       flag_threshold = c(-0.1, 0.1)
     )
     expect_equal(observed$fbmi, 1L)